refactor(gateway): relocate reply conflict handling
